Make POST tools/{mediapackageid}/editor.json Asynchronous
When exiting the video editor by choosing the "publish" workflow and
hitting "save & continue", the UI is blocked until the publish task is
created. On productive systems, this can take up to minutes depending
on the performance of the backend and on the amount of data it holds.

This patch changes the behavior of the ToolsEndpoint. Instead of waiting
for the tasks to be created and responding afterwards, an immediate
response is issued while the task is being created in the background.
This way, the UI can return to the list of recordings and the user can
go on editing other events.
To indicate that the task is not yet running, the workflow state of
the corresponding event is set to "PENDING". If any error occurs during
task creation, the workflow state is set to "FAILED".