MH-12595 Improve Publications Usability
This commit changes the design of the publications view to match the
style guide as well as it improves the usability.

In order to match the style guide, the following changes had to be done
to the "publications" tab in the event details:

- Adjust size of icons
- Add description text
- Add lines between publications

To improve the usability, the following changes were done:

- Add listprovider configuration to configure label, description, icon,
  whether to be shown or not and the order in the list for each
  publication
- Change some texts to improve understanding

So now, admins can configure the appearance of the "publications" box in
the event details to show only relevant publications in a meaningful
order with an explanatory description. However, if nothing is
configured, all values fall back to their defaults.

The "order" and "hidden" options also apply to the tooltip in the events
table when hovering over "yes" in the "published" column.