MH-12613 New WorkflowOperationHandler 'create-event'
This patch adds a new WOH to create new events from an existing one.
This is useful when you have a recording with - let's  say 3 - different
presentations and you want to create 3 events where each event only
includes one presentation. With this WOH, you would copy the original
event 2 times and cut each of the 3 events differently to only include
one presentation, each.

It would be nice to have this functionality in the video editor, but as
a first approach, this WOH is a good trade-off between effort and
usability. Also, this WOH is useful for creating test data.

This WOH also provides the possibility to select the media package
elements to be copied by tags and flavors. Tags can be changed for the
newly created media package elements using a "+" and "-" syntax.